+++ src/python/author.py
@@ -1,5 +1,0 @@
-class Author:
-    def __init__(self, name, email) -> None:
-        self.name = name
-        self.email = email
-        self.details = {}